app/requirements.txt
flask
redis
PyMySQL
Flask-SQLAlchemy
mysql-connector-python